# support per-org metrics_active for scraping by prometheus