closes #1159