@woodsaj 2018-12-14
no noticeable change in benchmark performance.  The +/- 2% is due to my desktop doing other things.

```
$ benchcmp master.txt perOrgActiveMetrics.txt
benchmark                                     old ns/op     new ns/op     delta
BenchmarkProcessMetricDataUniqueMetrics-4     35.1          34.4          -1.99%
BenchmarkProcessMetricDataSameMetric-4        34.4          34.8          +1.16%

benchmark                                     old allocs     new allocs     delta
BenchmarkProcessMetricDataUniqueMetrics-4     0              0              +0.00%
BenchmarkProcessMetricDataSameMetric-4        0              0              +0.00%

benchmark                                     old bytes     new bytes     delta
BenchmarkProcessMetricDataUniqueMetrics-4     0             0             +0.00%
BenchmarkProcessMetricDataSameMetric-4        0             0             +0.00%
```